orchestrator/requirements.txt
# MaynDrive LLM-Supervised Automation System Requirements

# Core dependencies
pydantic>=2.0.0
requests>=2.28.0

# Endless Discovery System
Pillow>=10.0.0  # Image processing for screenshots
imagehash>=4.3.1  # Perceptual image hashing

# Operator Console (Web UI)
fastapi>=0.104.0  # Web framework
uvicorn[standard]>=0.24.0  # ASGI server
websockets>=12.0  # WebSocket support

# Development and testing dependencies (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0

# Note: Android SDK, ADB, and Docker should be installed separately
# This requirements.txt only covers Python dependencies